requirements.txt
pandas
numpy
statsmodels
seaborn
matplotlib
python-dotenv
openai
